requirements.txt
memory-profiler
torch
numpy
cupy-cuda11x
matplotlib
pandas
pytest
